Redesign of the world.
The World is now a gigantic sphere hull which is 100 sectors thick.
There is a center at sector 0,0,0 from which X, Y and Z axises are
expanding 131 sectors in each direction. The center has a nice
golden pyramid with a diamond core at the moment.
The player starts at the top of the sphere at Sector 1,130,1.